+++ model/dmp_modules/gridmap_mod.f
@@ -528,8 +528,8 @@
 
       IF(numPEs .GT. 1) THEN
 ! Calculate any load imbalance.
-         IMBALANCE = DBLE(maxval(ijksize3_all)- minval(ijksize3_all)) /    &
-            minval(ijksize3_all)*100.0
+         IMBALANCE = INT(DBLE(maxval(ijksize3_all)- minval(ijksize3_all)) /    &
+            minval(ijksize3_all)*100.0)
 ! Calculate potential speedup based on Amdahl's Law
          IF(IMBALANCE == 0)THEN
             AMDAHL_SPEEDUP='+Inf'
